requirements.txt
# TaleSeed API - Dependências necessárias
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
